Move Shift instances to exp
